+++ Domain/DTO/CustomerRequestDTO.cs
@@ -4,7 +4,7 @@
     {
         public string Name { get; set; }
         public string Email { get; set; }
-        public decimal Income { get; set; }
+        public int Income { get; set; }
 
         public CustomerRequestDto()
         {
